hw3/requirements.txt
accelerate==1.1.1
asttokens==3.0.0
bitsandbytes==0.44.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
diffusers==0.31.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@19dabe96362803fb0a9ae7073d03533966598b17
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
zipp==3.21.0
